move Error to lib.rs
